tornet/requirements.txt
stem==1.8.2
requests==2.31.0
pysocks==1.7.1
colorama==0.4.6
psutil==5.9.5
beautifulsoup4==4.12.2
matplotlib==3.7.2
networkx==3.1
pillow==10.0.0
scapy==2.5.0
flask==2.3.3
flask-socketio==5.3.5
cryptography==41.0.3
numpy==1.25.2
scikit-learn==1.3.0
selenium==4.11.2
tqdm==4.66.1
dnspython==2.4.2
python-whois==0.8.0
aiohttp==3.8.5
